Use postings to calculate label values cardinality
When label values cardinality endpoint is being called with at least one
matcher, we were passing that matcher to each `countLabelValueSeries`
call.

This, in turn called `PostingsForMatchers` with the provided matchers
plus an extra matcher fixing the label value.

This operation calculates the intersection of the postings for the
provided matchers with the label value postings. As you can see, the
"postings for the provided matchers" would be calculated for every
label value found, which is quite inefficient, and gets worse as more
matchers are provided.

This approach calculates those postings for the provided matchers in the
first place, and then intersects them with the postings of each one of
the labels, reducing drastically the amount of work performed.
Benchmarks show a 50% benefit, which makes sense, as only half of the
postings have to be retrieved now.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>